@@ -1,7 +1,7 @@
 {
   "name": "taskana",
   "version": "1.0.0",
-  "license": "Novatec gmbh",
+  "license": "Novatec GmbH",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -16,60 +16,61 @@
     "format:html": "prettier --write \"**/*.{html,}\"",
     "e2e-dev": "cypress install && cypress run",
     "e2e-dev:open": "cypress install && cypress open",
-    "e2e-standalone": "cypress install && cypress run --config-file cypress-standalone.json"
+    "e2e-standalone": "cypress install && cypress run --config-file cypress-standalone.json",
+    "ci": "yarn install --frozen-lockfile"
   },
   "private": true,
+  "resolutions": {
+    "webpack": "5.21.2"
+  },
   "dependencies": {
-    "@angular/animations": "9.1.12",
-    "@angular/cdk": "9.2.4",
-    "@angular/common": "9.1.12",
-    "@angular/core": "9.1.12",
-    "@angular/forms": "9.1.12",
-    "@angular/material": "9.2.4",
-    "@angular/platform-browser": "9.1.12",
-    "@angular/platform-browser-dynamic": "9.1.12",
-    "@angular/router": "9.1.12",
-    "@ngxs/devtools-plugin": "3.7.1",
+    "@angular/animations": "11.1.0",
+    "@angular/cdk": "11.1.0",
+    "@angular/common": "11.1.0",
+    "@angular/core": "11.1.0",
+    "@angular/forms": "11.1.0",
+    "@angular/material": "11.1.0",
+    "@angular/platform-browser": "11.1.0",
+    "@angular/platform-browser-dynamic": "11.1.0",
+    "@angular/router": "11.1.0",
+    "@circlon/angular-tree-component": "10.0.2",
     "@ngxs/store": "3.7.1",
-    "angular-svg-icon": "9.2.0",
-    "angular-tree-component": "8.5.6",
-    "bootstrap": "4.6.0",
+    "angular-svg-icon": "11.0.0",
     "chart.js": "2.9.4",
     "core-js": "3.8.3",
-    "file-saver": "1.3.3",
-    "jquery": "3.5.1",
+    "file-saver": "2.0.5",
     "ng2-charts": "2.4.2",
-    "ngx-bootstrap": "5.4.0",
-    "ngx-infinite-scroll": "9.0.0",
-    "popper.js": "1.16.1",
+    "ngx-bootstrap": "6.2.0",
+    "ngx-infinite-scroll": "10.0.1",
     "rxjs": "6.6.3",
-    "tslib": "1.14.1",
-    "zone.js": "0.10.3"
+    "tslib": "2.1.0",
+    "zone.js": "0.11.3"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "9.2.0",
-    "@angular-devkit/build-angular": "0.901.13",
-    "@angular/cli": "9.1.12",
-    "@angular/compiler": "9.1.12",
-    "@angular/compiler-cli": "9.1.12",
+    "@angular-builders/custom-webpack": "11.0.0",
+    "@angular-devkit/build-angular": "0.1101.1",
+    "@angular/cli": "11.1.1",
+    "@angular/compiler": "11.1.0",
+    "@angular/compiler-cli": "11.1.0",
+    "@ngxs/devtools-plugin": "3.7.1",
     "@types/jest": "26.0.20",
     "@types/node": "14.14.22",
-    "@typescript-eslint/eslint-plugin": "2.34.0",
+    "@typescript-eslint/eslint-plugin": "4.14.0",
     "@typescript-eslint/eslint-plugin-tslint": "4.14.0",
-    "@typescript-eslint/parser": "2.34.0",
-    "compression-webpack-plugin": "4.0.0",
-    "cypress": "6.5.0",
+    "@typescript-eslint/parser": "4.14.0",
+    "compression-webpack-plugin": "7.1.2",
+    "cypress": "6.3.0",
     "cypress-intellij-reporter": "0.0.6",
-    "eslint": "6.8.0",
-    "eslint-config-prettier": "6.15.0",
+    "eslint": "7.18.0",
+    "eslint-config-prettier": "7.2.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-prettier": "3.3.1",
     "jest": "26.6.3",
     "jest-preset-angular": "8.3.2",
-    "node-sass": "4.14.1",
     "prettier": "2.2.1",
+    "sass": "1.32.5",
     "ts-jest": "26.4.4",
     "tslint": "6.1.3",
-    "typescript": "3.8.3"
+    "typescript": "4.1.3"
   }
 }
