@@ -1,16 +1,15 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { AngularSvgIconModule } from 'angular-svg-icon';
 import { RouterModule } from '@angular/router';
-import { TreeModule } from 'angular-tree-component';
-import { AlertModule, TypeaheadModule } from 'ngx-bootstrap';
+import { TreeModule } from '@circlon/angular-tree-component';
+import { AlertModule } from 'ngx-bootstrap/alert';
+import { TypeaheadModule } from 'ngx-bootstrap/typeahead';
 import { BsDatepickerModule } from 'ngx-bootstrap/datepicker';
 
 import { AccordionModule } from 'ngx-bootstrap/accordion';
-import { WorkbasketService } from 'app/shared/services/workbasket/workbasket.service';
-import { ClassificationsService } from 'app/shared/services/classifications/classifications.service';
 
 /**
  * Components
@@ -44,7 +43,6 @@ import { DateTimeZonePipe } from './pipes/date-time-zone.pipe';
  * Services
  */
 import { HttpClientInterceptor } from './services/http-client-interceptor/http-client-interceptor.service';
-import { AccessIdsService } from './services/access-ids/access-ids.service';
 import { ToastComponent } from './components/toast/toast.component';
 import { DialogPopUpComponent } from './components/popup/dialog-pop-up.component';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -72,7 +70,7 @@ const MODULES = [
   MatDialogModule,
   MatButtonModule,
   RouterModule,
-  TreeModule.forRoot(),
+  TreeModule,
   MatAutocompleteModule
 ];
 
@@ -120,11 +118,8 @@ const DECLARATIONS = [
       provide: HTTP_INTERCEPTORS,
       useClass: HttpClientInterceptor,
       multi: true
-    },
-    AccessIdsService,
-    ClassificationsService,
-    WorkbasketService
+    }
   ],
-  entryComponents: [ToastComponent, DialogPopUpComponent]
+  entryComponents: [DialogPopUpComponent, ToastComponent]
 })
 export class SharedModule {}
