@@ -1,7 +1,5 @@
 context('TASKANA Login', () => {
-  
-    it('should login depending on current configuration', () => {
-      cy.loginAs('admin');
-    });
-  
+  it('should login depending on current configuration', () => {
+    cy.loginAs('admin');
+  });
 });
