@@ -2,7 +2,8 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AngularSvgIconModule } from 'angular-svg-icon';
-import { AlertModule, TypeaheadModule } from 'ngx-bootstrap';
+import { AlertModule } from 'ngx-bootstrap/alert';
+import { TypeaheadModule } from 'ngx-bootstrap/typeahead';
 import { SharedModule } from 'app/shared/shared.module';
 import { InfiniteScrollModule } from 'ngx-infinite-scroll';
 
@@ -114,7 +115,6 @@ const DECLARATIONS = [
     WorkbasketDefinitionService,
     ClassificationCategoriesService,
     ImportExportService
-  ],
-  entryComponents: []
+  ]
 })
 export class AdministrationModule {}
