@@ -19,12 +19,10 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { RequestInProgressService } from '../../../shared/services/request-in-progress/request-in-progress.service';
 
 const getDomainFn = jest.fn().mockReturnValue(true);
-const domainServiceMock = jest.fn().mockImplementation(
-  (): Partial<DomainService> => ({
-    getDomains: getDomainFn,
-    getSelectedDomain: jest.fn().mockReturnValue(of('A'))
-  })
-);
+const domainServiceMock: Partial<DomainService> = {
+  getDomains: getDomainFn,
+  getSelectedDomain: jest.fn().mockReturnValue(of('A'))
+};
 
 @Component({ selector: 'taskana-administration-import-export', template: '' })
 class ImportExportStub {
@@ -44,11 +42,7 @@ class FilterStub {
   @Input() isExpanded = false;
 }
 
-const requestInProgressServiceSpy = jest.fn().mockImplementation(
-  (): Partial<RequestInProgressService> => ({
-    setRequestInProgress: jest.fn().mockReturnValue(of())
-  })
-);
+const requestInProgressServiceSpy = jest.fn().mockImplementation(() => jest.fn().mockReturnValue(of()));
 
 describe('WorkbasketListToolbarComponent', () => {
   let fixture: ComponentFixture<WorkbasketListToolbarComponent>;
@@ -70,8 +64,8 @@ describe('WorkbasketListToolbarComponent', () => {
       ],
       declarations: [WorkbasketListToolbarComponent, ImportExportStub, SortStub, FilterStub],
       providers: [
-        { provide: DomainService, useClass: domainServiceMock },
-        { provide: RequestInProgressService, useClass: requestInProgressServiceSpy },
+        { provide: DomainService, useValue: domainServiceMock },
+        { provide: RequestInProgressService, useValue: requestInProgressServiceSpy },
         WorkbasketService
       ]
     }).compileComponents();
