@@ -5,18 +5,15 @@ import { BrowserModule, By } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TaskanaEngineService } from '../../services/taskana-engine/taskana-engine.service';
 import { TaskanaEngineServiceMock } from '../../services/taskana-engine/taskana-engine.mock.service';
-import { of } from 'rxjs/internal/observable/of';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
-const TaskanaEngingeServiceSpy = jest.fn().mockImplementation(
-  (): Partial<TaskanaEngineServiceMock> => ({
-    hasRole: jest.fn().mockReturnValue(of()),
-    isHistoryProviderEnabled: jest.fn().mockReturnValue(of())
-  })
-);
+jest.mock('angular-svg-icon');
 
 @Component({ selector: 'svg-icon', template: '' })
-class SvgIconStub {}
+class SvgIconStub {
+  @Input() src;
+  @Input() matTooltip;
+}
 
 describe('UserInformationComponent', () => {
   let component: UserInformationComponent;
@@ -27,7 +24,7 @@ describe('UserInformationComponent', () => {
     TestBed.configureTestingModule({
       declarations: [UserInformationComponent, SvgIconStub],
       imports: [BrowserModule, HttpClientTestingModule, BrowserAnimationsModule],
-      providers: [{ provide: TaskanaEngineService, useClass: TaskanaEngingeServiceSpy }]
+      providers: [{ provide: TaskanaEngineService, useClass: TaskanaEngineServiceMock }]
     }).compileComponents();
   }));
 
